pyproject.toml
[project]
name = "toddlertransducer"
version = "0.1.0"
description = ""
authors = [
    {name = "Edward Parry",email = "red_parry@yahoo.co.uk"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-vlc",
    "flask",
    "mfrc522;  platform_machine == 'aarch64'",
    "RPi.GPIO;  platform_machine == 'aarch64'",
    "spidev;  platform_machine == 'aarch64'"
]

[tool.pylint.main]
source-roots = "./src/*"

[tool.pylint.format]
max-line-length = 120

[tool.poetry]
packages = [{include = "toddler_transducer", from = "src"}]

[tool.poetry.group.dev.dependencies]
black = "^25.9.0"
flake8 = "^7.3.0"
isort = "^6.0.1"
mypy = "^1.18.2"
pylint = "^3.3.8"
pytest = "8.4.2"


[tool.poetry.group.deployment.dependencies]
waitress = "^3.0.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
launch_toddler_transducer_web_app = "toddler_transducer.web_ui.launch:launch_toddler_transducer_web_app"
launch_dev_toddler_transducer_web_app = "toddler_transducer.web_ui.launch:launch_dev_toddler_transducer_web_app"
launch_toddler_transducer = "toddler_transducer.__main__:main"
